+++ cardano-cli/ChangeLog.md
@@ -21,6 +21,8 @@
   - stake-address delegation-certificate
   - stake-address deregistration-certificate
 
+- Add `--socket-path` CLI option for CLI commands that use `CARDANO_NODE_SOCKET_PATH` ([PR 4910](https://github.com/input-output-hk/cardano-node/pull/4910))
+
 ### Bugs
 
 - Allow reading signing keys from a pipe ([PR 4342](https://github.com/input-output-hk/cardano-node/pull/4342))